+++ samples/echo-source.lua
@@ -1,12 +1,13 @@
 --[[
 	Records a user's microphone and echos it back to them.
 
-	This uses the included QueueableSource object, which may still have issues.
+	This version uses vanilla LOVE sources, which is not ideal.
 ]]
 
 -- Alias love-microphone as microphone
 local microphone = require("love-microphone")
-local device, source
+
+local device
 
 function love.load()
 	-- Report the name of the microphone we're going to use
@@ -15,13 +16,9 @@ function love.load()
 	-- Open the default microphone device with default quality and 100ms latency
 	device = microphone.openDevice(nil, nil, 0.1)
 
-	-- Create a new QueueableSource to echo our audio
-	source = microphone.newQueueableSource()
-
 	-- Register our local callback
 	device:setDataCallback(function(device, data)
-		source:queue(data)
-		source:play()
+		love.audio.newSource(data):play()
 	end)
 
 	-- Start recording
@@ -29,6 +26,6 @@ function love.load()
 end
 
 -- Add microphone polling to our update loop
-function love.update()
+function love.update(dt)
 	device:poll()
 end